Feature #9, refactoring outlook calendar interactions to separate class
