+++ TFSTasksInOutlook/Properties/Settings.settings
@@ -8,5 +8,8 @@
     <Setting Name="Fake" Type="System.String" Scope="User">
       <Value Profile="(Default)" />
     </Setting>
+    <Setting Name="ShowOnlyWorkItemIdInCalendar" Type="System.Boolean" Scope="User">
+      <Value Profile="(Default)">False</Value>
+    </Setting>
   </Settings>
 </SettingsFile>
