+++ TFSTasksInOutlook/app.config
@@ -11,17 +11,20 @@
 	<startup><supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.6" /></startup>
 	<applicationSettings>
 		<TFSTasksInOutlook.Properties.Settings>
-			<setting name="TimesheetReportUrl" serializeAs="String">
-				<value>http://w0141db05/Reports_INSTANCE_2/Pages/Report.aspx?ItemPath=%2fTfsReports%2fDefaultCollection%2fAdministrative+Tasks%2fTimeSheet+Report</value>
-			</setting>
-		</TFSTasksInOutlook.Properties.Settings>
+   <setting name="TimesheetReportUrl" serializeAs="String">
+    <value>http://w0141db05/Reports_INSTANCE_2/Pages/Report.aspx?ItemPath=%2fTfsReports%2fDefaultCollection%2fAdministrative+Tasks%2fTimeSheet+Report</value>
+   </setting>
+  </TFSTasksInOutlook.Properties.Settings>
 	</applicationSettings>
 	<userSettings>
 		<TFSTasksInOutlook.Properties.Settings>
-			<setting name="Fake" serializeAs="String">
-				<value />
-			</setting>
-		</TFSTasksInOutlook.Properties.Settings>
+   <setting name="Fake" serializeAs="String">
+    <value />
+   </setting>
+   <setting name="ShowOnlyWorkItemIdInCalendar" serializeAs="String">
+    <value>False</value>
+   </setting>
+  </TFSTasksInOutlook.Properties.Settings>
 	</userSettings>
   <runtime>
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
